Add SaucerSensor class and electrical scheme for plant unit
- Implemented SaucerSensor class with methods to initialize and check if the saucer is full.
- Added initial electrical scheme for the plant unit in KiCad.

+++ src/plant/HumiditySensor.h
@@ -7,7 +7,7 @@ class HumiditySensor {
  public:
   HumiditySensor(int p);
   void begin();
-  int getHumidity();
+  float getHumidity();
   float getVoltage();
   int getValue();
